Fix incorrect eytzinger array in example
The previous eytzinger array was `4 2 5 1 6 3 7 8`. This is incorrect as it shows 6 as being 2's child, implying 6 is greater than 4... 3 being 5's child implying 3 is greater than 4... etc

To maintain the example with left-right-right traversal, I changed the array to be length 7 so there's exactly 4 leaf nodes with 4 as the root. An 8-element eytzinger array would have 5 as root and have a different traversal pattern.